kuzmichlawfirm
The Kuzmich Law Firm, P.C. is a state-wide practice that is located in the historic center of Old Town Lewisville, Texas. The city of Lewisville is a growing and prosperous suburb near the Dallas/Fort Worth (DFW) area of North Texas. The firm’s attorneys practice in state and federal court throughout the entire state of Texas.